Formatter is basically complete, Emoticons plugin updated
The Formatter is functional and has two gateways: format and revert.
format will set up the lexer with individual formatting patterns and parse
it, returning HTML-formatted text. revert is the opposite of format; it
takes HTML-formatted text and reverts it to a format that can be edited.
The Formatter is not yet extensible.

Posts are still stored in the database as HTML, and must be run through
revert in order to be edited. This has been done because it is the most
efficient way. Plugins, however, may choose to apply formatting before the
posts are displayed, so a user can turn off a plugin and its formatting
won't be corrupted. The Emoticons plugin now uses this method.

Some of the emoticons in the Emoticons plugin are tweaked.

Signed-off-by: Toby <toby@esotalk.com>