Various changes and fixes to the formatter
The previous commit introduced the "display" entry point for the formatter.
This has been removed as it does not work as hoped for. The Emoticons
plugin used this entry point to convert emoticon text entities into images
by means of regular expression search and replace, but this meant that it
converted emoticons which were inside code blocks. The only proper way of
preventing this and similar issues would be to use a lexer to determine
when the formatting should and should not take place, but this would defeat
the purpose of not using a lexer at display-level in the first place.

Thus, unfortunately, plugins must now use the "format" entry point and must
also include a "revert" entry point. See the Emoticons plugin for example.

A few minor formatter bugs are fixed.

Signed-off-by: Toby <toby@esotalk.com>